requirements.txt
absl-py==2.2.1
attrs==25.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.0
cycler==0.12.1
darkdetect==0.8.0
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.55.3
fsspec==2024.10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.4.5
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mediapipe==0.10.20
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opt_einsum==3.4.0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==4.25.6
psutil==6.1.0
py-cpuinfo==9.0.0
pycparser==2.22
pyparsing==3.2.0
pypiwin32==223
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
PySide6-Fluent-Widgets==1.7.6
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
sentencepiece==0.2.0
shiboken6==6.8.1
six==1.17.0
sounddevice==0.5.1
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zipp==3.21.0
zstandard==0.23.0

pyjwt~=2.10.1
flask~=3.1.0
